Extract common functions
