+++ app/api/management/commands/update_pdb_redo.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import time
 from urllib.parse import urljoin
 import requests
@@ -7,19 +6,7 @@ from django.core.management.base import BaseCommand
 from api.models import PdbEntry, RefinedModel, RefinedModelMethod, RefinedModelSource
 from api.dataPaths import URL_PDB_REDO
 from api.utils import save_json, updateRefinedModel
-
-
-HTTP_TIMEOUT = 15
-
-logger = logging.getLogger(__name__)
-logging.basicConfig(
-    level=logging.INFO, format="[%(asctime)s] %(levelname)s: %(message)s"
-)
-
-
-def log_info(message):
-    logger.info(message)
-    print(message)
+from .update_utils import log_info, save_entries, log_progress, HTTP_TIMEOUT
 
 
 class Command(BaseCommand):
@@ -37,7 +24,9 @@ class Command(BaseCommand):
         success, not_found = get_refined_model_pdb_redo(pdb_entries_list)
         log_info("Success: " + str(len(success)))
         log_info("Not found: " + str(len(not_found)))
-        save_entries(success, not_found)
+        save_entries(
+            success, not_found, "pdb_redo", save_json, "/data/pdb_redo_entries"
+        )
         update_pdb_redo_entries(success, not_found)
         log_info("** Finished updating PDB Redo entries **")
 
@@ -61,23 +50,10 @@ def get_refined_model_pdb_redo(list):
             success.append(pdb_id)
         else:
             not_found.append(pdb_id)
-        if time.time() - start_time >= 30:
-            progress = (index + 1) / len(list) * 100
-            log_info(
-                f"Progress: {progress:.2f}% - Success: {len(success)} - Not found: {len(not_found)}"
-            )
-            start_time = time.time()
+        log_progress(index, len(list), success, not_found, start_time)
         time.sleep(1)  # Avoiding DDoS
 
     return success, not_found
-
-
-def save_entries(success, not_found):
-    json = {"success": success, "not_found": not_found}
-    timestamp = time.strftime("%Y%m%d-%H%M%S")
-    filename = f"pdb_redo_{timestamp}.json"
-    save_json(json, "/data/pdb_redo_entries", filename)
-    log_info(f"Saved to {filename}")
 
 
 def get_refined_models():
